add unit test, add user_role in models, modify login required.

+++ src/surfing_penguin/API/User.py
@@ -1,10 +1,10 @@
 """routes.py: Each function in this file indicates a web page (HTML page)."""
 from src.surfing_penguin import surfing_penguin
-from src.surfing_penguin.extensions import login_manager
 from src.surfing_penguin.routes import api
 from flask_restplus import Resource, fields
+from functools import wraps
 from src.surfing_penguin.db_interface import UserFunctions
-from flask_login import login_user, logout_user, current_user, login_required
+from flask_login import login_user, logout_user, current_user
 
 
 # TODO: separate expected and returned api models
@@ -32,6 +32,21 @@ api_show_user_and_time = api.model("show_user_and_time_model", {
         'username': fields.String,
         'last_seen': fields.DateTime
     })
+
+
+def login_required(role="ANY"):
+    def wrapper(fn):
+        @wraps(fn)
+        def decorated_view(*args, **kwargs):
+            if not current_user.is_authenticated:
+                return error_message("not login")
+
+            if ((current_user.user_role not in role.split(',')) and (role !=
+                                                                     "ANY")):
+                return error_message("role")
+            return fn(*args, **kwargs)
+        return decorated_view
+    return wrapper
 
 
 @surfing_penguin.before_request
@@ -72,7 +87,7 @@ class login(Resource):
         name = api.payload['username']
         password = api.payload['password']
         if UserFunctions.get_user(name) is None:
-            return {'messages': "User not found"}
+            return error_message("User not found")
         if UserFunctions.check_password(name, password) is False:
             return {'messages': "Wrong passwd"}
         user = UserFunctions.get_user(name)
@@ -102,11 +117,13 @@ class show_users(Resource):
 class delete_user(Resource):
     @api.marshal_with(api_return_message)
     @api.expect(api_get_user)
-    @login_required
+    @login_required(role='ANY')
     def post(self):
         name = api.payload['username']
+        if current_user.user_role != 'admin' and current_user.username != name:
+            return error_message("role")
         if UserFunctions.search_user(name) is False:
-            return {'messages': "User not found"}
+            return error_message("User not found")
         UserFunctions.delete_user(name)
         return {'messages': "User {} deleted".format(name)}
 
@@ -118,12 +135,16 @@ class search_user(Resource):
     def post(self):
         search_name = api.payload['username']
         if UserFunctions.get_user(search_name) is None:
-            return {'messages': "user does not exist"}
+            return error_message("User not found")
         return UserFunctions.get_user(search_name)
 
 
-@login_manager.unauthorized_handler
 @api.marshal_with(api_return_message)
-def unauthorized():
-    # TODO: Flash this message somewhere.
+def error_message(message):
+    if message == "not login":
+        return {'messages': "Please Login First"}
+    if message == "role":
+        return {'messages': "You don't have permission!"}
+    if message == "User not found":
+        return {'messages': "User not found"}
     return {'messages': "Please Login First"}
